requirements.txt
pika==1.2.0
google-cloud-storage==1.42.3